split value.rs
